refactor: register the consume methods in the appropriate aspects
