+++ scopes/dependencies/pnpm/index.ts
@@ -1,3 +1,3 @@
-export { PnpmPackageManager } from './pnpm.package-manager';
 export type { PnpmMain } from './pnpm.main.runtime';
-export { PnpmAspect } from './pnpm.aspect';
+export type { PnpmUI } from './pnpm.ui.runtime';
+export { PnpmAspect as default, PnpmAspect } from './pnpm.aspect';
